backend/Facenet_Face_Recognition_App/pyproject.toml
[tool.poetry]
name = "facenet-face-recognition-app"
version = "0.1.0"
description = ""
authors = ["Samuvel6826 <samuvel6826@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
torch = "^2.2.0"
torchvision = "^0.17.0"
facenet-pytorch = "^2.6.0"
numpy = "^1.24.0"
matplotlib = "^3.9.2"
opencv-python = "^4.10.0.84"
scikit-learn = "^1.5.2"
python-dotenv = "^1.0.1"
firebase-admin = "^6.5.0"
flask = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
